requirements.txt
numpy
matplotlib
tqdm
tensorboard
torch
torchvision
jaxtyping
jax[cpu]
optax
equinox
pyyaml
